30 May 2022 2

+++ Inventories/Controllers/TransferController.cs
@@ -55,6 +55,8 @@ namespace Inventories.Controllers
             HttpCookie cookieObj = Request.Cookies["User"];
             int UserId = Int32.Parse(cookieObj["UserId"]);
             tblTransfer Data = new tblTransfer();
+            tblCheckinItem Update = new tblCheckinItem();
+            tblAdjustmentItem AUpdate = new tblAdjustmentItem();
             string TNumber = HeadData[0].TransferNumber;
             if (HeadData.FirstOrDefault().TransferId == 0)
             {
@@ -101,9 +103,29 @@ namespace Inventories.Controllers
                 {
                     if (First)
                     {
+                        Update = new tblCheckinItem();
+                        AUpdate = new tblAdjustmentItem();
                         Item.TransferId = Data.TransferId;
                         Item.CreatedBy = UserId;
                         Item.CreatedDate = DateTime.Now;
+                        Item.ItemUsedQuantity = 0;
+                        Item.ItemNetQuantity = Item.ItemQuantity;
+                        if (Item.VNumber[0] == 'C')
+                        {
+                            Update = DB.tblCheckinItems.Where(x => x.CheckinItemId == Item.CheckinItemId).FirstOrDefault();
+                            Update.ItemUsedQuantity = Update.ItemUsedQuantity + Item.ItemQuantity;
+                            Update.ItemNetQuantity = Update.ItemQuantity - Update.ItemUsedQuantity;
+                            DB.Entry(Update);
+                            DB.SaveChanges();
+                        }
+                        if (Item.VNumber[0] == 'A')
+                        {
+                            AUpdate = DB.tblAdjustmentItems.Where(x => x.AdjustmentItemId == Item.CheckinItemId).FirstOrDefault();
+                            AUpdate.ItemUsedQuantity = AUpdate.ItemUsedQuantity + Item.ItemQuantity;
+                            AUpdate.ItemNetQuantity = AUpdate.ItemQuantity - AUpdate.ItemUsedQuantity;
+                            DB.Entry(AUpdate);
+                            DB.SaveChanges();
+                        }
                         DB.tblTransferItems.Add(Item);
                     }
                     else
@@ -235,9 +257,50 @@ namespace Inventories.Controllers
             List<tblTransferItem> Data1 = new List<tblTransferItem>();
             Data = DB.tblTransfers.Where(x => x.TransferId == Id).FirstOrDefault();
             Data1 = DB.tblTransferItems.Where(x => x.TransferId == Id).ToList();
-            DB.tblTransferItems.RemoveRange(Data1);
-            DB.tblTransfers.Remove(Data);
-            DB.SaveChanges();
+            tblCheckinItem Update = new tblCheckinItem();
+            tblAdjustmentItem AUpdate = new tblAdjustmentItem();
+            bool Check = false;
+            foreach (tblTransferItem item in Data1)
+            {
+                if (item.ItemUsedQuantity > 0)
+                {
+                    Check = true;
+                    break;
+                }
+                
+            }
+            if (Check == false)
+            {
+                foreach (tblTransferItem item in Data1)
+                {
+                    Update = new tblCheckinItem();
+                    AUpdate = new tblAdjustmentItem();
+                    if (item.VNumber[0] == 'C')
+                    {
+                        Update = DB.tblCheckinItems.Where(x => x.CheckinItemId == item.CheckinItemId).FirstOrDefault();
+                        Update.ItemUsedQuantity -= item.ItemQuantity;
+                        Update.ItemNetQuantity = Update.ItemQuantity - Update.ItemUsedQuantity;
+                        DB.Entry(Update);
+                        DB.SaveChanges();
+                    }
+                    if (item.VNumber[0] == 'A')
+                    {
+                        AUpdate = DB.tblAdjustmentItems.Where(x => x.AdjustmentItemId == item.CheckinItemId).FirstOrDefault();
+                        AUpdate.ItemUsedQuantity -= item.ItemQuantity;
+                        AUpdate.ItemNetQuantity = AUpdate.ItemQuantity - AUpdate.ItemUsedQuantity;
+                        DB.Entry(AUpdate);
+                        DB.SaveChanges();
+                    }
+
+                }
+                DB.tblTransferItems.RemoveRange(Data1);
+                DB.tblTransfers.Remove(Data);
+                DB.SaveChanges();
+            }
+            else
+            {
+                return Json(2);
+            }
 
             return Json(1);
         }
@@ -249,6 +312,60 @@ namespace Inventories.Controllers
             try
             {
                 allsearch = DB.tblItems.Where(x => x.Name.StartsWith(search)).ToList();
+
+            }
+            catch (Exception ex)
+            {
+                ViewBag.Error = ex.Message;
+                Console.WriteLine("Error" + ex.Message);
+            }
+
+            return new JsonResult { Data = allsearch, JsonRequestBehavior = JsonRequestBehavior.AllowGet };
+        }
+
+
+        public JsonResult GetCheckinList(int ItemId, int WarehouseId = 0)
+        {
+            List<GetTCheckinList_Result> allsearch = null;
+            DB.Configuration.ProxyCreationEnabled = false;
+            try
+            {
+                allsearch = DB.GetTCheckinList(ItemId, WarehouseId).ToList();
+
+            }
+            catch (Exception ex)
+            {
+                ViewBag.Error = ex.Message;
+                Console.WriteLine("Error" + ex.Message);
+            }
+
+            return new JsonResult { Data = allsearch, JsonRequestBehavior = JsonRequestBehavior.AllowGet };
+        }
+
+        public JsonResult GetCheckinItemData(int CheckinItemId, string Number)
+        {
+            GetCheckinItemData_Result3 allsearch = null;
+            GetTransferItemData_Result3 Tallsearch = null;
+            GetAdjustmentItemData_Result2 Aallsearch = null;
+            DB.Configuration.ProxyCreationEnabled = false;
+            try
+            {
+                if (Number[0] == 'T')
+                {
+                    Tallsearch = DB.GetTransferItemData(CheckinItemId).FirstOrDefault();
+                    return new JsonResult { Data = Tallsearch, JsonRequestBehavior = JsonRequestBehavior.AllowGet };
+                }
+                if (Number[0] == 'C')
+                {
+                    allsearch = DB.GetCheckinItemData(CheckinItemId).FirstOrDefault();
+                    return new JsonResult { Data = allsearch, JsonRequestBehavior = JsonRequestBehavior.AllowGet };
+                }
+                if (Number[0] == 'A')
+                {
+                    Aallsearch = DB.GetAdjustmentItemData(CheckinItemId).FirstOrDefault();
+                    return new JsonResult { Data = Aallsearch, JsonRequestBehavior = JsonRequestBehavior.AllowGet };
+                }
+
 
             }
             catch (Exception ex)
